+++ resources/views/user/view-ticket.blade.php
@@ -12,7 +12,17 @@
             <h5 class="modal-title">Rate the Helpdesk Service</h5>
             <button type="button" class="btn-close" data-bs-dismiss="modal" aria-label="Close"></button>
             </div>
-            <div class="modal-body">
+            <div class="modal-body" id="ticket_rate_body">
+            <div class="alert alert-info" role="alert">
+                <div class="d-flex">
+                    <div>
+                        <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round" class="icon alert-icon icon-2"><path d="M3 12a9 9 0 1 0 18 0a9 9 0 0 0 -18 0"></path><path d="M12 9h.01"></path><path d="M11 12h1v4h1"></path></svg>
+                    </div>
+                    <div>
+                    This Ticket isn't rated yet.
+                    </div>
+                </div>
+            </div>
                 <form id="rateTicketForm" class="needs-validation" novalidate>
                     <div class="mb-3">
                         <label class="form-label">Ticket Rating <span class="text-danger">*</span></label>
@@ -144,7 +154,7 @@
                     </div>
             </div>
                 <div class="modal-footer">
-                    <button type="submit" id="rateTicketFormSubmit" class="btn btn-primary">Submit</button>
+                    <button type="submit" id="rateTicketFormSubmit" class="btn btn-primary" disabled>Submit</button>
                 </div>
             </form>
         </div>